Create bold label component and move the shared component from sub-folder
