+++ app/components/shared/cardWithSoundWave/CardWithSoundWaveAudioComponent.js
@@ -1,20 +1,32 @@
 import React from 'react';
 import { View, Text, StyleSheet } from 'react-native';
-import { Card } from 'react-native-paper';
 
-import AudioWaveButtonComponent from '../audioWaveButton/AudioWaveButtonComponent';
-import CardWithSoundWaveInfoComponent from './CardWithSoundWaveInfoComponent';
-import color from '../../../themes/color';
-import {cardElevation} from '../../../constants/component_constant';
-import {normalFontSize, smallFontSize} from '../../../utils/font_size_util';
+import AudioWaveButtonComponent from '../AudioWaveButtonComponent';
+import {smallFontSize} from '../../../utils/font_size_util';
+import componentUtil from '../../../utils/component_util';
 
 const CardWithSoundWaveAudioComponent = (props) => {
   return (
     <View>
-      <AudioWaveButtonComponent containerStyle={{ width: 48, height: 48, borderWidth: 0, position: 'absolute', top: -20, zIndex: 10 }} />
-      <Text style={{textAlign: 'right', fontSize: smallFontSize()}}>02:30</Text>
+      <AudioWaveButtonComponent containerStyle={styles.btn} />
+      <Text style={styles.label}>02:30</Text>
     </View>
   )
 }
+
+const styles = StyleSheet.create({
+  btn: {
+    borderWidth: 0,
+    height: componentUtil.pressableItemSize(),
+    width: componentUtil.pressableItemSize(),
+    position: 'absolute',
+    top: -20,
+    zIndex: 10
+  },
+  label: {
+    fontSize: smallFontSize(),
+    textAlign: 'right',
+  }
+});
 
 export default CardWithSoundWaveAudioComponent;
